+++ software/healthcheck.c
@@ -163,7 +163,7 @@ bool inmHealthCheckAddBit(bool evenBit, bool oddBit, bool even, uint8_t addr) {
     inmTotalBits++;
     if(inmDebug && (inmTotalBits & 0xfffff) == 0) {
         fprintf(stderr, "Generated %llu bits.  %s to use data.  Estimated entropy per bit: %f, estimated K: %f\n",
-            inmTotalBits, inmHealthCheckOkToUseData()? "OK" : "NOT OK", inmHealthCheckEstimateEntropyPerBit(),
+            (long long)inmTotalBits, inmHealthCheckOkToUseData()? "OK" : "NOT OK", inmHealthCheckEstimateEntropyPerBit(),
             inmHealthCheckEstimateK());
         fprintf(stderr, "num1s:%f%%, even misfires:%f%%, odd misfires:%f%%\n",
             inmTotalOnes*100.0/(inmTotalZeros + inmTotalOnes),
